printer-api/requirements.txt
flask==3.1.*
flask-cors==5.*
pycups==2.0.*
